Bugfix: Database should keep field order in update broadcasts.
The reasoning here is that its possible to override a previous field update in the same packet by either sending the same field twice.
An example if when this WOULDN'T be a mistake, is if someone wanted to send two molecular fields that had one or more fields of intersection.