Unittests: Improve debug out for expectMany.
To improve the debug output a little, we actually try to match the failing
datagram against the most similar datagram.
If the channels are the same as an expected, we're going to guess that is
what we want; otherwise we try the first datagram with matching length;
if nothing else, then we always try the first datagram. (We infer that the
order of the datagrams is in EXPECTED order of receipt even if their may
be datagram race conditions.)