Owner views are also subject to interest visibility
In a nutshell, a client can have two views of an object:
a DistributedObject instance and a DistributedObjectOV

The owner view instance (DistObjectOV) is created when the
client receives CLIENT_ENTER_OBJECT_REQUIRED[_OTHER]_OWNER
(i.e. as soon as STATESERVER_OBJECT_SET_OWNER) is dispatched.

The "regular" view, on the other hand, only exists while the
client has interest in the object's location.

For example, a game may give to the client the ownership of a
DistributedShip object upon the avatar's login. That causes
DistributedShipOV to generate. With that view, the client
can hear some important updates: setMasts, setArmor and even
send ownsend updates.

When the ship is deployed, the server moves it to a certain
location so that the client may find it with an interest
and generate a DistributedShip instance which loads the actual
model with controls.

Toontown implements none of this and this commit shouldn't
affect it. Pirates Online, on the other hand, relies a lot
on this.

To support my claim, I'd like to point out that:

1 - Panda3D implements self.doId2ownerView and self.doId2do
as two distinct maps.

2 - cConnectionRepository.cxx, in handle_update_field_owner,
attempts to update both views. This goes to show that both
instances do coexist.

Furthermore, many more evidences are present in
Pirates Online / TLOPO code. Please contact me if
you'd like to see them.