ClientAgent: Remove SSL/TLS support
This is in preparation for the migration to libuv, which does not have any
builtin support for SSL/TLS. SSL/TLS is still possible via Stunnel/HAProxy
and both of these solutions support the PROXY protocol to pass through the
client IP address information.

In fact, it would be my personal recommendation that users of Astron do
their SSL/TLS this way - SSL/TLS is complicated and getting it set up
securely involves a lot of configuration, and the builtin CA SSL/TLS
configuration did not support many of the more advanced options
(e.g. enabled cipher suites and minimum TLS version) that professional
game security teams would want to be able to set. Furthermore, doing the
SSL/TLS out-of-process improves the parallelism/pipelining of a game
cluster (at the expense of a small latency hit - less than 1ms) and allows
SSL/TLS termination nodes to be scaled independently of CAs as necessary.

The SSL/TLS test support code is left intact; maybe someday fate will have
it that first-class SSL/TLS capability is added back to Astron, but for now
the Astron project is not large enough that it can guarantee correct and
fully-functional SSL/TLS capability and that responsibility should be given
to a project with many more security-focused eyeballs on it instead.